refactor: service 패키지 내 interface와 impl 패키지 분리
- service 패키지 내 interface와 impl 패키지를 분리하여 코드 구조 개선
