+++ src/main/java/gift/service/impl/KakaoMessageServiceImpl.java
@@ -1,5 +1,6 @@
-package gift.service;
+package gift.service.impl;
 
+import gift.service.KakaoMessageService;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
